+++ doc/changelog.rst
@@ -41,6 +41,7 @@ Localization
 Admin
 ~~~~~
 
+- Fix product "Add new image" link
 - Fix content block styles that are styled by id
 - Add Orders section to product detail page
 - Add `admin_product_section` provide to make product detail extendable